[feat] 293 report api

See merge request s13-fintech-finance-sub1/S13P21B108!168

+++ walletslot-backend/src/main/java/com/ssafy/b108/walletslot/backend/domain/slot/repository/AccountSlotRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface AccountSlotRepository extends JpaRepository<AccountSlot, Integer> {
+public interface AccountSlotRepository extends JpaRepository<AccountSlot, Long> {
 
     Optional<List<AccountSlot>> findByAccount(Account account);
     void deleteByAccount(Account account);
